requirements.txt
fastapi>=0.110.0
jinja2
python-multipart
uvicorn
asyncio
pytest
requests
python-dotenv

sqlalchemy>=2.0.27
cityhash
fastembed
redis>=5.0.2
rq
rq-dashboard
pyarango>=2.1.1

numpy>=1.26.4
pandas>=2.2.2

Faker
markdown

google-generativeai>=0.7.2

transformers
ctransformers[cuda]
sentence-transformers

qdrant-client>=1.12.0